@@ -1,6 +1,6 @@
 package sk.greate43.bucketdrops.holder;
 
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.View;
 import android.widget.Button;
 
